rename Toggle to ToggleButton
