+++ .changeset/serious-beans-explain.md
@@ -1,0 +1,5 @@
+---
+"@optiaxiom/react": minor
+---
+
+rename Toggle to ToggleButton
