+++ apps/docs/demos/toggle-button/disabled-usage/App.tsx
@@ -1,10 +1,10 @@
-import { Toggle, Tooltip } from "@optiaxiom/react";
+import { ToggleButton, Tooltip } from "@optiaxiom/react";
 import { IconLayoutSidebar } from "@tabler/icons-react";
 
 export function App() {
   return (
     <Tooltip content="Disabled toggle demo">
-      <Toggle
+      <ToggleButton
         aria-label="Toggle sidebar"
         disabled
         icon={<IconLayoutSidebar />}
