+++ apps/docs/pages/components/toggle-button.mdx
@@ -1,15 +1,15 @@
 import { Demo } from "@/components/demo";
-import { ControlledUsage, DisabledUsage, Usage } from "@/demos/toggle";
+import { ControlledUsage, DisabledUsage, Usage } from "@/demos/toggle-button";
 
-# Toggle
+# ToggleButton
 
-`Toggle` component represents a button that can be toggled on or off.
+`ToggleButton` component represents a button that can be toggled on or off.
 
 ## Documentation
 
 ### Usage
 
-`Toggle` extends the [Button](./button/) component.
+`ToggleButton` extends the [Button](./button/) component.
 
 <Demo component={Usage} />
 
@@ -29,7 +29,7 @@ For accessibility please make sure to include text that explains why the button 
 
 ## Props
 
-<ax-props-table component="Toggle" extends="Button" />
+<ax-props-table component="ToggleButton" extends="Button" />
 
 ## Changelog
