+++ apps/storybook/src/forms/ToggleButton.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Toggle } from "@optiaxiom/react";
+import { ToggleButton } from "@optiaxiom/react";
 import { IconLayoutSidebar } from "@tabler/icons-react";
 
 export default {
@@ -17,10 +17,10 @@ export default {
   args: {
     icon: "layout-sidebar",
   },
-  component: Toggle,
-} as Meta<typeof Toggle>;
+  component: ToggleButton,
+} as Meta<typeof ToggleButton>;
 
-type Story = StoryObj<typeof Toggle>;
+type Story = StoryObj<typeof ToggleButton>;
 
 export const Basic: Story = {};
