+++ packages/react/src/index.ts
@@ -83,7 +83,7 @@ export * from "./toast";
 export * from "./toast-action";
 export * from "./toast-provider";
 export * from "./toast-title";
-export * from "./toggle";
+export * from "./toggle-button";
 export * from "./tokens";
 export * from "./tooltip";
 export * from "./tooltip-content";
