--- packages/react/src/toggle/index.ts
@@ -1,1 +1,0 @@
-export * from "./Toggle";
