+++ packages/web-components/package.json
@@ -97,7 +97,7 @@
     "./ToastAction": "./dist/components/ToastAction.js",
     "./ToastProvider": "./dist/components/ToastProvider.js",
     "./ToastTitle": "./dist/components/ToastTitle.js",
-    "./Toggle": "./dist/components/Toggle.js",
+    "./ToggleButton": "./dist/components/ToggleButton.js",
     "./Tooltip": "./dist/components/Tooltip.js",
     "./TooltipContent": "./dist/components/TooltipContent.js",
     "./TooltipRoot": "./dist/components/TooltipRoot.js",
